refactor: standardise implementation of non-official attributes
